refactor: move OnboardingLayout
