+++ packages/desktop/components/index.js
@@ -11,7 +11,6 @@ export { default as AccountSwitcher } from './AccountSwitcher.svelte'
 export { default as ContactAddressCard } from './ContactAddressCard.svelte'
 export { default as ContactCard } from './ContactCard.svelte'
 export { default as ContactMetadataTable } from './ContactMetadataTable.svelte'
-export { default as OnboardingLayout } from './OnboardingLayout.svelte'
 export { default as NetworkCard } from './NetworkCard.svelte'
 export { default as NodeListTable } from './NodeListTable.svelte'
 export { default as Proposals } from './Proposals.svelte'
