+++ packages/desktop/views/components/OnboardingLayout.svelte
