+++ packages/desktop/views/components/index.js
@@ -1,0 +1,1 @@
+export { default as OnboardingLayout } from './OnboardingLayout.svelte'
